backend/requirements.txt
# Backend Requirements

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic[email]>=2.5.0

# Already in main project
# agents
# openai
# pymupdf
# groq
# google-generativeai
# reportlab
# python-docx
# python-pptx
# pyttsx3
# requests
# beautifulsoup4
